Added cactus
If you are above cactus, you can't go into spectator. But I didn't add to the side, as you could be in a desert, and cactus is less than a block wide, and I don't know how to check that specifically. Feel free to add it if you know how